pyproject.toml
[project]
name = "project-x"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.6.4",
    "ollama>=0.5.1",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
    "twilio>=9.7.0",
    "uvicorn>=0.35.0",
]
